+++ XADLZSSHandle.m
@@ -72,7 +72,7 @@
 		else
 		{
 			matchlength=length;
-			matchoffset=pos-offset;
+			matchoffset=(int)(pos-offset);
 		}
 	}
